Cluster analysis script changes
File path concatenation chenged to use fullfile (to be executable on Linuc and Windows).
Added a plotFlag variable to switch figure output on or off.
Fixed some code warnings (variable preallocation) in the initialization section.
Added an initialization section (and an accompanying cleanup section) to setup the parallel processing pool for execution on the bwUniCluster3.0 (according to the bwUniCluster wiki). In the repo version, the code is commented out.
Added a small section at the beginning to support batch processing (currently the whole file list is a single batch, but it's now possible to execute in several smaller batches).
Added time measurements for dataset processing.
Fixed some code warnings (unused variables) inside the main loop; all unused lines are just commented out, not removed.
Several functionality changes (as follows), all of which have been commented in the code.
The image data is now imported as floating-point data. This was necessary because not all included functionality works on uint data. The conversion is currently done by simply casting. It could be changed to use im2double, but then intensity values might be less interpretable. Note that this increases memory requirements by roughly a factor of four regarding the image data variables. Note that this change also has a significant influence on the following process steps, because the image data can become negative after background removal. This further influences Otsu threholding. To return closer to uint processing, the image data may be clipped after the background removal step.
Histogram computation has been changed to use modern functions (the old hist function is deprecated). The result is very close, but not identical to before.
Some result variables were supposedly not set as intended. This has been corrected, with detailed comments.
A whole lot of temporary variables from inside the main loop are now cleaned before saving the results. In parallel processing, these variables are actually loop-private and don't exist afterwards, but in serial processing, they had taken up a whole lot of unnecessary space in the result files.
Some whitespace cosmetics.