More changes to cluster analysis script
The default processing mode has been changed to serial; the reasoning for this is that parallel on the compute cluster is extremely badly performing (for currently unknown reasons), and on a local computer, parallel processing may quickly raise memory issues. Serial processing on a local computer is performing ok, and the performance issues on the compute cluster seem to be less grave in serial mode (but that also may be purely coincidental in the performed tests).
Nevertheless, the removal of the parallel cluster and pool variables at the end of the processing loop has been taken into the code; it doesn't do anything wrong if they don't exist.
Fixed buggy behaviour when specifying empty values for the S5P_nucClustVolume and S5P_nucClustVolume variables.
All temporary loop-internal variables are now correctly cleaned before saving the workspace after the loop.
The save file name variable is now actually used when saving the workspace.